Format leaderboard announcement time in minutes and seconds, add commas to score
In the AnnounceLeaderboardEntry method, the time associated with a leaderboard score was previously displayed in seconds only. This commit modifies this to display in a more user-friendly minutes and seconds format. The change will improve usability by making leaderboard times easier to read and compare.

Additionally, scores are now properly formatted with commas (e.g. 1000 -> 1,000) which makes it look nicer.